moved functions to common
